+++ NEWS.md
@@ -1,5 +1,7 @@
 # shinymde 0.1.0.9000
 
+* Users now have a reset button at the "Input Data" stage that may be useful in case of mind changes. 
+
 * `run_app` was renamed `launch_app` to avoid name conflicts for example with `golem::run_app`. 
 
 * A simple welcome message was added to ensure that users have something visible as they select a dataset. It also serves as a form of credit to the author(s).